Added nextjs api routes
There will be error with new user registration. I believe that that is the only api endpoint that sends data as `multipart/form-data` instead of json.

The default is not to use the nextjs api endpoint. Possible feature to implement: Switch to using it by setting `getAPI_URL` to return it when offline.