Bug Fix #240
Current issue is caused when category is not set, hence the default option is not selected. I think why the 1st option is not re-rendered is because the key is not changed, so react think that it is the same element. The key is only dependent on which position the diagnosis is in.