Finally figured out the issue
For some reason, the field value I get from the render prop in the Controller component is outdated, while the value from the watch method is updated. Swapping to that fixed the issue of the dropdown/selector switching to the wrong value.

However, that only fixed the gross_motor field. For the scoliosis and pallor fields, there was a defaultValue set in ChildVitalsFields that still kept setting the field to the wrong value. Removing that fixed the other 2 select fields.

The final issue is that after these 2 fixes, the select component jumps to the previous option before jumping back. The issue is that reset() is called right after patchVital, before the VitalsForm component has retrieved the updated data from backend. Thus, after resetting, the select component jumps to the previous (wrong) value. When the VitalsForm component finishes retrieving the updated data, it then jumps back.